src/utils/date-helpers/date-helpers.ts
export const formatDate = (date: Date) => {
  const rawDate = new Date(date);
  const month = rawDate.getMonth() + 1;
  const day = rawDate.getDate();

  return `${day}/${month}`;
};

export const formatTime = (dateRaw: number) => {
  const rawDate = new Date(dateRaw);
  const hours = rawDate.getHours().toString().padStart(2, '0');
  const minutes = rawDate.getMinutes().toString().padStart(2, '0');
  return `${hours}:${minutes}`;
};

export const isToday = (someDate: Date) => {
  const today = new Date();
  return (
    someDate.getDate() === today.getDate() &&
    someDate.getMonth() === today.getMonth() &&
    someDate.getFullYear() === today.getFullYear()
  );
};

export const isYesterday = (someDate: Date) => {
  const yesterday = new Date();
  yesterday.setDate(yesterday.getDate() - 1);
  return (
    someDate.getDate() === yesterday.getDate() &&
    someDate.getMonth() === yesterday.getMonth() &&
    someDate.getFullYear() === yesterday.getFullYear()
  );
};
